+++ src/Object/init.lua
@@ -1,3 +1,5 @@
+export type Object = { [string]: any? }
+
 return {
 	assign = require(script.assign),
 	entries = require(script.entries),